// src/store/types/TTermSlice.ts
export type DROPDOWN_MASTER = {
  insurance_company: [];
  SMOKE: TDropdown[];
  GENDER: TDropdown[];
  ANNUAL_INCOME: TDropdown[];
  AGE: TDropdown[];
  SUM_ASSURED: TDropdown[];
  PAY_MODE_DATA: TDropdown[];
  COVERAGE_UPTO: TDropdown[];
  PAY_UPTO: TDropdown[];
  SORT_BY: TDropdown[];
  MONTHLY_PAYOUT: TDropdown[];
  OCCUPATION_DATA: TDropdown[];
  OCCUPATION_DATA_GODIGIT: TDropdown[];
  QUALIFICATION_DATA: TDropdown[];
  INCOME_PERIOD: TDropdown[];
  PLAN_TYPE: TDropdown[];
  EDUCATION_MASTER: TDropdown[];
  MAX_ANNUAL_INCOME: TDropdown[];
  MAX_SUM_ASSURED: TDropdown[];
  myProf: TDropdown[];
  orgType: TDropdown[];
  indsType: TDropdown[];
  buyingPolicy: TDropdown[];
  insuranceComp: TDropdown[];

  state: TDropdown[];
  PREGNANCY_WEEKS: TDropdown[];
  purposeTravel: TDropdown[];
  typeResidence: TDropdown[];
  modeTravel: TDropdown[];
  AADHAR_OPTIONS_DATA: TDropdown[];
  IT_PROOF_MASTER: TDropdown[];
  IDENTITY_PROOF_MASTER: TDropdown[];
  BENEFICIARY_RELATIONSHIP: TDropdown[];
  TRUSTEE_TYPE: TDropdown[];
  eia: TDropdown[];
  vaccine: TDropdown[];
  occupationData: TDropdown[];
  genderData: TDropdown[];
  maritalStatusData: TDropdown[];
  nomineeRelationshipData: TDropdown[];
  annualIncomeData: TDropdown[];
  heightFeetData: TDropdown[];
  heightInchesData: TDropdown[];
  tobacco: TDropdown[];
  alcohol: TDropdown[];
  addressProf: TDropdown[];
  ageProf: TDropdown[];
  incomeProf: TDropdown[];
  identityProf: TDropdown[];
  income_tax_proof: TDropdown[];
  country: TDropdown[];
  HDFC_CKYC_OCCUPATION: TDropdown[];
  HDFC_MEDICAL_DESIGNATION: TDropdown[];
  NATURE_OF_WORK: TDropdown[];
  COMPANY_BELONGS: TDropdown[];
  SELECT_ANY_SIMULTANEOUS: TDropdown[];
  HDFC_COUNTRIES: TDropdown[];
  HOBBIES: TDropdown[];
  MAY_UNDERGO_MEDICAL_EXAMINATION: TDropdown[];
  CONSUME_ALCOHOL: TDropdown[];
  TOBACCO_PRODUCT: TDropdown[];
  RELATION_OF_LIFE_TO_BE_ASSURED: TDropdown[];
  HDFC_DISEASE: TDropdown[];
  AGE_AT_DIAGNOSIS: TDropdown[];
  HDFC_ALIVE: TDropdown[];
  HDFC_CURRENT_AGE: TDropdown[];
  HDFC_HEIGHT: TDropdown[];
  HDFC_WEIGHT: TDropdown[];
  have_you_suffered_from_one: TDropdown[];
  have_you_suffered_from_two: TDropdown[];
  have_you_suffered_from_three: TDropdown[];
  have_you_suffered_from_four: TDropdown[];
  have_you_suffered_from_five: TDropdown[];
};

export type TTermQuoteFieldData = {
  payMode: string;
  sortBy: string;
  payUpto: string;
  monthlyPayout: string;
  coverageUpto: string;
  sumAssured: string;
  planType: "1" | "2" | "3" | "4";
  income_period: string;
};

export type TTermProposerData = {
  premium_loader: boolean;
  is_proposer_is_insured_member: boolean;
  full_name: TFieldData;
  dob: TFieldData;
  gender: TFieldData;
  mobile: TFieldData;
  email: TFieldData;
  aadhar_number: TFieldData;
  emergency_number: TFieldData;
  marital_status: TFieldData;
  annual_income: TFieldData;
  education: TFieldData;
  occupation: TFieldData;
  occupation_desc: TFieldData;
  org_name: TFieldData;
  org_type: TFieldData;
  org_description: TFieldData;
  industry_type: TFieldData;
  industry_type_desc: TFieldData;
  my_profession: TFieldData;
  org_type_desc: TFieldData;
  obj_of_buying_policy: TFieldData;
  pan_no: TFieldData;
  sum_assured: TFieldData;
  nationality: "indian" | "nri";
  diabetic: boolean;
  pincode: TFieldData;
  qualification: TFieldData;
  insured_name: TFieldData;
  insured_mobile: TFieldData;
  insured_email: TFieldData;
  insured_gender: TFieldData;
  insured_dob: TFieldData;
  insured_occupation: TFieldData;
  insured_qualification: TFieldData;
  insured_pincode: TFieldData;
  corporate_email_id: TFieldData;
  isSalariedDiscount: TFieldData;
  isSalariedConsent: TFieldData;
  lifeMileStoneDiscount: TFieldData;
  lifeMileStoneType: TFieldData;
  lifeMileStonedate: TFieldData;
  housewifePayerStatus: boolean;
  housewifePayerRelation: TFieldData;
  first_time_buyer: boolean;
  existing_customer: boolean;
};

export type TTermSlice = {
  CUS_TYPE: boolean;
  DROPDOWN_DATA: DROPDOWN_MASTER;
  ADD_FORM: TTermAddForm;
  ADD_FORM_RESPONSE: TTermAddFormResponse;
  NO_QUOTE_REASONS: TTermNoQuoteReason[];
  QUOTE_FIELDS_DATA: TTermQuoteFieldData;
  SELECTED_QUOTE_DATA: TSelectedQuote;
  RIDERS_BENEFITS: TRidersBenefits;
  PROPOSER_DATA: TTermProposerData;
  QUOTE_LOAD_FIRST_TIME: boolean;
  LOADER: boolean;
  redirection_url: string;
  REDIRECTED: boolean;
};

export type TSelectedQuote = {
  transaction_id: string;
  appno: string;
  documentURL: string;
  premiumDetails: {
    modal_factor: string;
    product_desc: string;
    income_period: string;
    term: string;
    pay_term: string;
    sum_assured: number;
    sum_insured?: number;
    annual_premium?: number;
    wop_rider_premium: number;
    wop_rider_gst: number;
    wopPremium: number;
    ci_rider_premium: number;
    ci_rider_gst: number;
    ciPremium: number;
    acc_rider_premium: number;
    acc_rider_gst: number;
    accPremium: number;
    income: string;
    premium: number;
    riders: {
      TF001: TSingleRider;
      TF002: TSingleRider;
      TF003: TSingleRider;
      TF004: TSingleRider;
      TF005: TSingleRider;
    };
    premiumAmount: number;
    Total_basePremium: number;
    Total_taxAmount: number;
    taxAmount: string;
    finalPremium: number;
    annunalTargetPremium: number;
    product_id: number;
    product_code: string;
    cover_pay_upto: string;
    frequency: string;
    smoker: string;
    cover_upto: string;
    policy_covers: number;
  };
  companyDetails: {
    company_code: string;
    name: string;
    short_desc: string;
    logo: string;
    claim_ratio: string;
  };
  SpecialFeatureLists: any[];
  productDetailLists: null | any;
  buyOnlineLink: {
    method: string;
    buy_online_link: string;
    redirection_type: string;
  };
  productDetails: {
    id: number;
    product_name: string;
    brochure: string;
    product_code: string;
    rate_calc_method: string;
    proposal_method: string;
    tax_calc_method: string;
    discount_calc_method: string;
    rating: number;
    product_desc: string;
    buy_online_code: string;
    special_feature_codes: string;
    prem_calc_action: string;
    tax_calc_action: string;
    discount_calc_action: string;
    term_form_validation: {
      modal_factor: number;
      min_sum_assured: number;
      max_sum_assured: number;
      min_age: number;
      max_age: number;
      min_term: number;
      max_term: number;
      min_p_cess_age: number;
      max_p_cess_age: number;
    };
    frequency: string;
    CompanyProductMaps: {
      company_code: string;
    };
  };
  sumAssured: string;
  QuoteNo: string;
};